# workload: allow columnar data generation (alloc -73%, MB/s +60%)