--- a/pkg/ccl/utilccl/sampledataccl/bankdata.go
+++ b/pkg/ccl/utilccl/sampledataccl/bankdata.go
@@ -53,7 +53,7 @@ func toBackup(t testing.TB, data workload.Table, dir string, chunkBytes int64) (
 	var stmts bytes.Buffer
 	fmt.Fprintf(&stmts, "CREATE TABLE %s %s;\n", data.Name, data.Schema)
 	for rowIdx := 0; rowIdx < data.InitialRows.NumBatches; rowIdx++ {
-		for _, row := range data.InitialRows.Batch(rowIdx) {
+		for _, row := range data.InitialRows.BatchRows(rowIdx) {
 			rowBatch := strings.Join(workload.StringTuple(row), `,`)
 			fmt.Fprintf(&stmts, "INSERT INTO %s VALUES (%s);\n", data.Name, rowBatch)
 		}