A workload.Generator previously could specifiy its initial table data
using one of two methods. One outputs a single row as `[]interface{}`
and the other is a batch of `[][]interface{}`. This commit switches the
latter to be a columar batch, allowing selected performance-critical
workload generators (tpcc, bank) to generate initial table data with
dramatically reduced allocations.

The single row option still exists and will continue to be used by most
workload.Generators, as it's much simpler. An adaptor from the columnar
batch to the old row oriented `[][]interface{}` batch is included for
ease of use in non-performance sensitive consumers of initial table
data. IMPORT, on the other hand, is switched to using the columnar
batches directly.

Some of the savings here is reusing batches, but the majority comes from
`[]interface{}` requiring that everything assinged to it be on the heap,
while something like `[]int` doesn't. We also get some amount of speedup
in initial table data consumers from fewer type switches.

This could also be accomplished without being columnar, but doing the
minimum non-columnar thing to get the same results will require
reinventing most of the work that the columnar does. Plus it will almost
certainly be convenient for columnar exec benchmarking to have a common
format.

Benchmark results:

    name                               old time/op    new time/op    delta
    InitialData/tpcc/warehouses=1-8       552ms ± 1%     454ms ± 0%  -17.81%  (p=0.008 n=5+5)
    InitialData/bank/rows=1000-8          461µs ± 0%     332µs ± 1%  -28.10%  (p=0.008 n=5+5)
    WriteCSVRows-8                       15.0µs ± 1%    16.8µs ± 1%  +12.44%  (p=0.008 n=5+5)
    CSVRowsReader-8                      17.8µs ± 1%    19.3µs ± 1%   +8.41%  (p=0.008 n=5+5)

    name                               old speed      new speed      delta
    InitialData/tpcc/warehouses=1-8     128MB/s ± 1%   209MB/s ± 0%  +63.22%  (p=0.008 n=5+5)
    InitialData/bank/rows=1000-8        223MB/s ± 0%   350MB/s ± 1%  +56.83%  (p=0.008 n=5+5)
    WriteCSVRows-8                      118MB/s ± 2%   107MB/s ± 3%   -9.43%  (p=0.008 n=5+5)
    CSVRowsReader-8                     101MB/s ± 4%    92MB/s ± 5%   -8.31%  (p=0.016 n=5+5)

    name                               old alloc/op   new alloc/op   delta
    InitialData/tpcc/warehouses=1-8       246MB ± 0%      75MB ± 0%  -69.61%  (p=0.008 n=5+5)
    InitialData/bank/rows=1000-8          232kB ± 0%      19kB ± 0%  -91.75%  (p=0.008 n=5+5)
    WriteCSVRows-8                       5.60kB ± 0%    5.70kB ± 0%   +1.71%  (p=0.016 n=4+5)
    CSVRowsReader-8                      7.35kB ± 1%    7.39kB ± 0%     ~     (p=0.667 n=5+4)

    name                               old allocs/op  new allocs/op  delta
    InitialData/tpcc/warehouses=1-8       5.60M ± 0%     1.48M ± 0%  -73.56%  (p=0.008 n=5+5)
    InitialData/bank/rows=1000-8          6.00k ± 0%     1.02k ± 0%  -83.01%  (p=0.008 n=5+5)
    WriteCSVRows-8                         48.0 ± 0%      50.0 ± 0%   +4.17%  (p=0.008 n=5+5)
    CSVRowsReader-8                        54.0 ± 0%      55.0 ± 0%   +1.85%  (p=0.008 n=5+5)

We could probably speed the CSV stuff back up with some followup work,
but (a) the overall `fixtures import` benchmark is still faster and (b)
we're about to switch to dt's import magic that skips the CSV roundtrip.

Current `fixtures import`

    name                               old time/op    new time/op    delta
    ImportFixture/tpcc/warehouses=1-8     3.35s ± 2%     3.29s ± 3%     ~     (p=0.310 n=5+5)

    name                               old speed      new speed      delta
    ImportFixture/tpcc/warehouses=1-8  26.4MB/s ± 2%  26.9MB/s ± 3%     ~     (p=0.310 n=5+5)

    name                               old alloc/op   new alloc/op   delta
    ImportFixture/tpcc/warehouses=1-8    3.09GB ± 0%    2.92GB ± 0%   -5.60%  (p=0.008 n=5+5)

    name                               old allocs/op  new allocs/op  delta
    ImportFixture/tpcc/warehouses=1-8     32.1M ± 0%     28.0M ± 0%  -12.86%  (p=0.008 n=5+5)

Magic `fixtures import` that skips CSV roundtrip:

    name                               old time/op    new time/op    delta
    ImportFixture/tpcc/warehouses=1-8     2.88s ± 3%     2.86s ± 4%     ~     (p=0.548 n=5+5)

    name                               old speed      new speed      delta
    ImportFixture/tpcc/warehouses=1-8  30.7MB/s ± 3%  30.9MB/s ± 4%     ~     (p=0.548 n=5+5)

    name                               old alloc/op   new alloc/op   delta
    ImportFixture/tpcc/warehouses=1-8    2.87GB ± 0%    2.73GB ± 0%   -4.97%  (p=0.008 n=5+5)

    name                               old allocs/op  new allocs/op  delta
    ImportFixture/tpcc/warehouses=1-8     26.1M ± 0%     21.5M ± 0%  -17.75%  (p=0.008 n=5+5)

Touches #34809

Release note: None