@danhhz 2019-03-25
Okay @jordanlewis this is as far as I got on Friday. There's a bit of a performance regression, so I haven't bothered updated the commit message or PR description yet. This is a bummer and there's definitely a bunch of places I can optimize, but I'm torn between how huge and scary this already is vs having a window between landing this and the followup perf work to really get all the benefits of this. Note that even though the csv writing (which we still use for a little while longer) is slower, fixtures import tpcc already uses drastically fewer allocs and is overall a bit faster.

```
name                               old time/op    new time/op    delta
pkg:github.com/cockroachdb/cockroach/pkg/workload goos:darwin goarch:amd64
InitialData/tpcc/warehouses=1-8       628ms ± 2%     539ms ± 1%  -14.14%  (p=0.008 n=5+5)
InitialData/bank/rows=1000-8          572µs ±10%     752µs ± 2%  +31.40%  (p=0.008 n=5+5)
WriteCSVRows-8                       15.9µs ± 2%    18.9µs ± 2%  +18.93%  (p=0.008 n=5+5)
CSVRowsReader-8                      18.1µs ± 0%    20.7µs ± 1%  +14.05%  (p=0.016 n=4+5)
pkg:github.com/cockroachdb/cockroach/pkg/ccl/workloadccl goos:darwin goarch:amd64
ImportFixture/tpcc/warehouses=1-8     3.25s ± 1%     3.31s ± 0%   +1.83%  (p=0.016 n=5+4)

name                               old speed      new speed      delta
pkg:github.com/cockroachdb/cockroach/pkg/workload goos:darwin goarch:amd64
InitialData/tpcc/warehouses=1-8     113MB/s ± 2%   176MB/s ± 1%  +56.23%  (p=0.008 n=5+5)
InitialData/bank/rows=1000-8        180MB/s ± 9%   154MB/s ± 2%  -14.42%  (p=0.008 n=5+5)
WriteCSVRows-8                      114MB/s ± 3%    96MB/s ± 4%  -15.28%  (p=0.008 n=5+5)
CSVRowsReader-8                     101MB/s ± 4%    87MB/s ± 5%  -14.47%  (p=0.008 n=5+5)
pkg:github.com/cockroachdb/cockroach/pkg/ccl/workloadccl goos:darwin goarch:amd64
ImportFixture/tpcc/warehouses=1-8  27.2MB/s ± 1%  26.7MB/s ± 0%   -1.80%  (p=0.016 n=5+4)

name                               old alloc/op   new alloc/op   delta
pkg:github.com/cockroachdb/cockroach/pkg/workload goos:darwin goarch:amd64
InitialData/tpcc/warehouses=1-8       246MB ± 0%      75MB ± 0%  -69.51%  (p=0.008 n=5+5)
InitialData/bank/rows=1000-8          232kB ± 0%     347kB ± 0%  +49.41%  (p=0.016 n=5+4)
WriteCSVRows-8                       5.59kB ± 0%    7.74kB ± 0%  +38.44%  (p=0.016 n=5+4)
CSVRowsReader-8                      7.38kB ± 1%    7.38kB ± 0%     ~     (p=0.206 n=5+4)
pkg:github.com/cockroachdb/cockroach/pkg/ccl/workloadccl goos:darwin goarch:amd64
ImportFixture/tpcc/warehouses=1-8    2.56GB ± 0%    2.39GB ± 0%   -6.72%  (p=0.008 n=5+5)

name                               old allocs/op  new allocs/op  delta
pkg:github.com/cockroachdb/cockroach/pkg/workload goos:darwin goarch:amd64
InitialData/tpcc/warehouses=1-8       5.60M ± 0%     1.48M ± 0%  -73.56%  (p=0.008 n=5+5)
InitialData/bank/rows=1000-8          6.00k ± 0%     7.02k ± 0%  +16.97%  (p=0.008 n=5+5)
WriteCSVRows-8                         48.0 ± 0%      51.0 ± 0%   +6.25%  (p=0.008 n=5+5)
CSVRowsReader-8                        53.0 ± 0%      54.0 ± 0%   +1.89%  (p=0.008 n=5+5)
pkg:github.com/cockroachdb/cockroach/pkg/ccl/workloadccl goos:darwin goarch:amd64
ImportFixture/tpcc/warehouses=1-8     31.7M ± 0%     27.6M ± 0%  -13.02%  (p=0.016 n=5+4)
```