@jordanlewis 2019-03-27
Any initial guesses why it might be slower? I'll take a look soon.