@danhhz 2019-03-28
The fixtures import is almost certainly slower because CSV generation is slower. After 19.1.0 goes out the door, david is going to merge https://github.com/cockroachdb/cockroach/pull/36250, which switches fixtures import to bypass the csv by default and this is definitely faster with this PR. So even if we can't get the csv performance back, this whole problem goes away anyway.

As for why it's slower, I've tried digging into before/after profiles. My best guess so far is it's because most workload.Generators have a batch size of 1, which is the worst case for seeing columnar overhead.