@danhhz 2019-05-07
TestStoreRangeMergeRaftSnapshot was just modified by #37350 so I'm inclined to guess flake. Dunno what the story is with that TestEncodeBitArray panic but I doubt it's this PR.

bors r=jordanlewis